extracts MockDeviceCache
