+++ feature/amazon/src/test/java/com/revenuecat/purchases/amazon/mocks/MockDeviceCache.kt
@@ -1,0 +1,26 @@
+package com.revenuecat.purchases.amazon.mocks
+
+import android.content.SharedPreferences
+import com.revenuecat.purchases.common.caching.DeviceCache
+import org.json.JSONObject
+
+class MockDeviceCache(
+    preferences: SharedPreferences,
+    apiKey: String
+) : DeviceCache(preferences, apiKey) {
+
+    var stubCache = mutableMapOf<String, String>()
+
+    override fun getJSONObjectOrNull(key: String): JSONObject? {
+        return stubCache[key]?.let {
+            JSONObject(it)
+        }
+    }
+
+    override fun putString(
+        cacheKey: String,
+        value: String
+    ) {
+        stubCache[cacheKey] = value
+    }
+}
